Fix error encountered during large file transfers
In case the server sends data of size a constant multiple of the buffer
size, the client didn't take this scenario into account. Thus, even
though all the data was successfully transferred, there was no way the
client could've tell if the server is willing to send more data
afterwards. We fix this issue by making the server explicitly send an
empty message before it closes the connection from its end.

A better version of the protocol would be that instead of sending an
empty segment, the server sends some well defined message through which
the client can infer that the server no longer wants to send data.